requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
anvil-uplink==0.4.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
blinker==1.4
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
fonttools==4.25.0
frozenlist==1.3.1
future==0.18.2
google-auth==2.11.1
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.4
importlib-metadata==4.11.4
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
jupyter_client==7.3.5
jupyter_core==4.11.1
kiwisolver==1.4.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.2
munkres==1.1.4
nest-asyncio==1.5.5
numpy==1.22.4
oauthlib==3.2.1
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
ply==3.11
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pybase64==1.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pyu2f==0.1.5
pywin32==303
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==65.5.0
sip==6.6.2
six==1.16.0
stack-data==0.5.0
sympy==1.10.1
tensorboard==2.10.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
toml==0.10.2
torch==1.10.2
torchfile==0.1.0
torchvision==0.11.3
tornado==6.2
tqdm==4.64.0
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
ws4py==0.5.1
yarl==1.7.2
zipp==3.8.1
